SimplexInterpolation

Documentation for SimplexInterpolation

StarStats.InterpSimplex

StarStats.InterpSimplex

StarStats.SimplexInterpolant

StarStats.SimplexInterpolant

StarStats.StellarModelGrid

StarStats.barycentric_coords

StarStats.barycentric_coords!

StarStats.compute_distance_and_EEPs!

StarStats.compute_simplex_interpolation

StarStats.interpolation_info

StarStats.mesa_dataframe_loader

StarStats.InterpSimplex — Type

struct InterpSimplex

Structure to describe a simplex together with precomputed properties used to determine barycentric coordinates. In $N$ dimensions, given $N+1$ vertices on a simplex $\vec{x}_i$ and a point $\vec{a}$ for which we want to determine the barycentric coordinates $\alpha_i$, it should be satisfied that $\vec{a} = \sum_{i=1}^{N+1}\alpha_i \vec{x}_i$ together with the condition $\sum_{i=1}^{N+1}lpha_i = 1.$ This is equivalent to solving a linear system $A ec{lpha}= ec{b}$ where $A=[(\vec{x}_1-\vec{x}_N)\;\dots\;(\vec{x}_{N-1}-\vec{x}_N)],\qquad \vec{b}=\vec{a}-\vec{x}_N$ and $\vec{\alpha}$ contains th barycentric coordinates from $1$ to $(N-1)$. To efficiently determine barycentric coordinates we use StaticArrays with precomputed LU factorizations of the $A$ matrix. We also store the vertex $\vec{x}_N$ separately.

points: matrix containing the points with dimensions ndims x npoints

A_lu: factorized matrix in the linear system to determine the barycentric coordinate coefficients

end_point: Coordinates of $\vec{x}_N$, used to set up the value of $\vec{b}$ in the linear system when a point $\vec{a}$ is provided.

StarStats.InterpSimplex — Method

function InterpSimplex(points)

Construct a simplex based on the given points. points must be a matrix of dimensions ndims x npoints.

StarStats.SimplexInterpolant — Type

struct SimplexInterpolant{N,P,LU,E}

Structure containing information on a Delaunay triangulation to perform interpolation.

simplexes: Instances of Simplex for each of the simplexes in the triangulation

points: Points used for the triangulation. points must be a matrix of dimensions ndims x npoints.

points_indeces: For each simplex, contains the indeces of the points that compose it. Dimensions are ndims+1 x nsimplexes.

StarStats.SimplexInterpolant — Method

function SimplexInterpolant(points)

Constructor for a SimplexInterpolant based on a set of points. points must be a matrix of dimensions ndims x npoints.

StarStats.barycentric_coords! — Method

function barycentric_coords!(point, s::InterpSimplex{N,P,LU,E}, coords, bcache) where {N,P,LU,E}

Non-allocating calculation of barycenter coordinates. Results are stored in place in coords.

StarStats.barycentric_coords — Method

function barycentric_coords(point, s::InterpSimplex{N,P,LU,E}) where {N,P,LU,E}

Wrapper for barycentric_coords! that is non-mutating.

StarStats.compute_simplex_interpolation — Method

function compute_simplex_interpolation(info, values)

Given the results from the function interpolation_info and the values of a function evaluated in the points used to create the SimplexInterpolant, return the interpolated value. Returns NaN if no simplex contains the point.

StarStats.interpolation_info — Method

function interpolation_info(point, si::SimplexInterpolant{N,P,LU,E}) where {N,P,LU,E}

Iterate through all simplexes in the SimplexInterpolant si to find the simplex containing the given point. Returns the Barycentric coordinates of the point for the containing simplex, the indeces of the points that form the vertices of the simplex and the id of the simplex. If no simplex is found that contains the point zero values are returned for the indeces and coefficients.